requirements.txt
pandas
numpy
apache-airflow==2.8.1
